pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "youos"
version = "0.1.0"
description = "Your personal AI email copilot — learns your writing style from Gmail and drafts replies that sound like you."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic>=2.8,<3.0",
  "pydantic-settings>=2.4,<3.0",
  "pyyaml>=6.0.2,<7.0",
  "sqlmodel>=0.0.22,<0.1",
  "httpx>=0.27,<1.0",
  "python-dotenv>=1.0,<2.0",
  "typer>=0.12,<1.0",
  "rich>=13.0,<14.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "ruff>=0.6,<1.0",
  "mypy>=1.11,<2.0"
]

[project.scripts]
youos = "app.cli:app"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
